chore: split up Suggestion Component
