+++ client/package.json
@@ -11,6 +11,7 @@
 	],
 	"scripts": {
 		"dev": "vite",
+		"start": "vite",
 		"build": "vite build",
 		"lint": "eslint src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
 		"preview": "vite preview"
