+++ client/src/Components/Loader/Loader.jsx
@@ -1,0 +1,9 @@
+import loader from '../../assets/loader.gif';
+
+export default function Loader() {
+	return (
+		<div className="loader">
+			<img src={loader}></img>
+		</div>
+	);
+}
